Will now output as [xml] instead of [string]
The function when I found it was outputting the result to a formatted string.  Instead, I now take the encoding type and XML version, and add a XML declaration to the top of the document.  I then output the XML as type [xml], so it is ready to be worked with.